Updated Practice 3

+++ Practica4/consola.cpp
@@ -5,7 +5,7 @@
 // ***************************************************************************************
 void gotoxy(int x, int y) 
 { 
-    COORD pos = {x, y};
+    COORD pos = {static_cast<short int>(x), static_cast<short int>(y)};
     HANDLE output = GetStdHandle(STD_OUTPUT_HANDLE);
     SetConsoleCursorPosition(output, pos);
 }
